# Faster init with LoRA