Purpose: accelerate peft model initialisation with LoRA config.
see issue description in https://github.com/huggingface/peft/issues/871

The idea is to avoid costly calls of weight initialisation functions in CPU, when those weights are not necessary and will be replaced by the original weights anyways.

This PR initially changes only init for Linear layers. 

potential further improvements:
- apply to Embedding and Conv2d layers
- apply to compressed layer types, like Linear4bit, Linear8bit
- init LoRA modules (A, B) right on target device, not CPU