@poedator 2023-08-28
> I wonder if we can use [`skip_init`](https://pytorch.org/docs/stable/generated/torch.nn.utils.skip_init.html) or something else from PyTorch instead of monkey patching a hard-coded list of initialization methods.

well, I borrowed this idea [from GPTQ code](https://github.com/IST-DASLab/gptq/blob/2d65066eeb06a5c9ff5184d8cebdf33662c67faf/llama.py#L15). But I support your idea to avoid monkey patching.
Let me try this improvement - I will write back soon.