@poedator 2023-08-28
@BenjaminBossan !
`skip_init()` turned out not very convenient to apply from inside of a subclass. However, looking at its source code, I got an idea to use `meta` device. It works much faster than my original code, since meta tensors are faster than empty cpu tensors. Thanks for the suggestion!
To accomodate this change I also:
	- made `self.to(self.weight.device)` in LoraLayer.update_layer() conditional, to avoid moving loras to 'meta' device
	- removed  self.weight.requires_grad = False in `Linear.__init__()` for it is not relevant.   # lora.py:850

Now with Llama-2-7B application of Guanaco LoRA adapter takes 1.5 seconds vs 103 seconds originally.