@BenjaminBossan 2023-08-28
@poedator Great, thanks for experimenting. This looks like a very nice speedup indeed. I'm not 100% sure if we can remove all of that code, but let's see. Could you please run `make style` on your code? Then the code quality checks should pass and the unit tests should run.