@poedator 2023-08-28
I added couple more commits to pass device param to LoRA layer constructors to initialize LoRA adapters right at the target device. Apparently I need to do some more local testing ...
As a matter of introduction: I am a junior researcher @ Yandex, 2d name in this paper: https://arxiv.org/abs/2306.03078. We use HF packages daily and want them to get even better ).