@BenjaminBossan 2023-08-29
Thanks for applying the fixes and providing more context. 

Unfortunately, the tests are failing right now. The main reason seems to be that the code path when `init_lora_weights=False` is not properly implemented yet, resulting in the weights staying on meta device. This needs to be fixed and it might require a bit of fiddling, as the logic is a bit all over the place. It would probably help if we cleaned up the initialization logic, but that requires some non-trivial refactoring.

To proceed, would you be up for fixing the failing tests for the code path I mentioned?