@poedator 2023-08-29
@BenjaminBossan 
I updated the PR to catch up with #807 and also limited my changes to just the Linear layer init to prove the concept. 
Some of the tests with `init_lora_weights=False` are also failing locally (test_merge_layers_09 and 10). This is puzzling because they do not seem to involve LoRA for Linear. Nevertheless I want to see this PR coming thru, so let me try to fix it with your guidance.

pls be more specific on the failing tests. do you want me to update the tests code or to change some LoRA code?