@BenjaminBossan 2023-08-30
Thanks for the updates and the further investigation. Regarding the failing tests, the idea is that all existing tests should pass after the change. There could be tests that need adjustment, but I'm not aware of any. Anyway, for me, all the tests passed. As you can see, the CI passes too. So not sure why they failed for you.

I think the strategy you suggested, which is to start with one layer type first, is valid. We can do that and once that has proven to work well, we can tackle the others.

I did some further digging into the code and IIUC, I think the weight initialization happening in these two lines is completely unnecessary:

https://github.com/huggingface/peft/blob/0b2f950cc212dd8acda983760c0aba2bf4872a00/src/peft/tuners/lora/layer.py#L142
https://github.com/huggingface/peft/blob/0b2f950cc212dd8acda983760c0aba2bf4872a00/src/peft/tuners/lora/layer.py#L151

It creates a randomly initialized `self.weight` (and possibly `self.bias`) on the lora layer, which is then initialized with kaiming init twice! And later, it is replaced by the original weight of the target layer, so the whole initialization was unnecessary:

https://github.com/huggingface/peft/blob/0b2f950cc212dd8acda983760c0aba2bf4872a00/src/peft/tuners/lora/model.py#L213

Therefore, I think this step could be completely skipped. Since we're talking about the full sized weight here, not the smaller LoRA weights, this could take quite some time. The fix you proposed would create that weight on the meta device, but I wonder if we can rewrite the code to completely skip the initialization instead. If you want to explore that, feel free to do so, otherwise I'll take a look at it later. If you have some benchmark script to test the initialization speed, please share it.

> Also I may try to pass device argument to LoRAs inits so they are created on Cuda where they belong (still faster). But that may be better left to a separate PR.

Yes, that's indeed best left for a separate PR.