@poedator 2023-08-30
> @poedator I attempted to remove the redundant initializations completely in #887. If you have time to test potential speedups of that branch, it would be great.

Thanks for the initiative. Let me test #887 later today.